Restructure api doc folder
