+++ quackbox-api-docs/site/index.html
@@ -1,0 +1,20 @@
+<!DOCTYPE html>
+<html lang="en">
+  <head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <meta http-equiv="X-UA-Compatible" content="ie=edge">
+    <title>My Website</title>
+    <link rel="stylesheet" href="./style.css">
+    <link rel="icon" href="./favicon.ico" type="image/x-icon">
+  </head>
+  <body>
+    <main>
+        <h1>Quackbox API Documentation</h1>  
+    </main>
+    <button onclick="redirect('./docs-v1/index.html')">V1 docs</button>
+    <button onclick="redirect('./docs-v2/index.html')">V2 docs</button>
+    <script src="index.js">
+    </script>
+  </body>
+</html>
