+++ quackbox-api-docs/site/index.js
@@ -1,0 +1,3 @@
+function redirect(path) {
+    window.location.href = path;
+}
